Support classes with completely custom serialization
The library currently assumes that any SC_SERIALIZABLE with SC_WRITE_METHOD classes will start their writeObject implementation with a call to defaultWriteObject - but that does not have to be the case.

This commit adds a method to register a custom classdata handler function in order to allow consumers to support any kind of class. To preserve bcompat the existing behaviour is used as a fallback if no handler is registered.